+++ ICD.Connect.Routing.CrestronPro/DigitalMedia/DmMd128X128/DmMd128X128AdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmMd128X128
+{
+	[KrangSettings("DmMd128x128", typeof(DmMd128X128Adapter))]
+	public sealed class DmMd128X128AdapterSettings : AbstractCrestronBladeSwitchAdapterSettings
+	{
+	}
+}